refactor: move sol parsing to separate mod
